requirements_exact.txt
aistudio-sdk==0.3.8
altgraph==0.17.5
annotated-types==0.7.0
anyio==4.12.1
astor==0.8.1
attrdict==2.0.1
babel==2.17.0
bce-python-sdk==0.9.60
beautifulsoup4==4.14.3
blinker==1.9.0
cachetools==6.2.6
certifi==2026.1.4
chardet==5.2.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
colorlog==6.10.1
contourpy==1.3.2
cssselect==1.4.0
cssutils==2.11.1
cycler==0.12.1
Cython==3.2.4
decorator==5.2.1
et_xmlfile==2.0.0
exceptiongroup==1.3.1
filelock==3.20.3
fire==0.7.1
Flask==3.1.2
flask-babel==4.0.0
fonttools==4.61.1
fsspec==2026.1.0
future==1.0.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.3.5
idna==3.11
ImageIO==2.37.2
imagesize==1.4.1
imgaug==0.4.0
itsdangerous==2.2.0
Jinja2==3.1.6
keyboard==0.13.5
kiwisolver==1.4.9
lazy_loader==0.4
lmdb==1.7.5
lxml==6.0.2
MarkupSafe==3.0.3
matplotlib==3.10.8
modelscope==1.34.0
more-itertools==10.8.0
mss==10.1.0
networkx==3.4.2
Nuitka==2.8.10
numpy==1.26.4
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
opencv-python-headless==4.13.0.90
openpyxl==3.1.5
opt-einsum==3.3.0
ordered-set==4.1.0
packaging @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_packaging_1769093650/work
paddleocr==2.7.3
paddlepaddle==2.6.2
paddlex==3.4.1
pandas==2.3.3
pdf2docx==0.5.8
pefile==2024.8.26
pillow==12.1.0
premailer==3.10.0
prettytable==3.17.0
protobuf==3.20.2
psutil==7.2.2
py-cpuinfo==9.0.0
pyclipper==1.4.0
pycryptodome==3.23.0
pydantic==2.12.5
pydantic_core==2.41.5
pyinstaller==6.18.0
pyinstaller-hooks-contrib==2026.0
PyMuPDF==1.26.7
pyparsing==3.3.2
pypdfium2==5.3.0
PyQt6==6.10.2
PyQt6-Qt6==6.10.2
PyQt6_sip==13.11.0
PySide6==6.10.1
PySide6_Addons==6.10.1
PySide6_Essentials==6.10.1
python-bidi==0.6.7
python-dateutil==2.9.0.post0
python-docx==1.2.0
pytz==2025.2
pywin32-ctypes==0.2.3
PyYAML==6.0.2
RapidFuzz==3.14.3
rarfile==4.2
requests==2.32.5
ruamel.yaml==0.19.1
safetensors==0.7.0
scikit-image==0.25.2
scipy==1.15.3
shapely==2.1.2
shellingham==1.5.4
shiboken6==6.10.1
six==1.17.0
soupsieve==2.8.3
termcolor==3.3.0
tifffile==2025.5.10
tqdm==4.67.1
typer-slim==0.21.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
ujson==5.11.0
urllib3==2.6.3
visualdl==2.5.3
wcwidth==0.5.2
Werkzeug==3.1.5
zstandard==0.25.0
